requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
scipy>=1.7.0
